# crates/api-client/casiros.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CASIROS API",
    "description": "NASA/JPL-grade Financial Physics Engine & Multiverse Simulator",
    "contact": {
      "name": "CASIROS Contributors"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/evaluate": {
      "post": {
        "tags": [
          "handlers"
        ],
        "summary": "Evaluates a causality graph with fixed inputs.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Evaluation succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or evaluation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/simulate": {
      "post": {
        "tags": [
          "handlers"
        ],
        "summary": "Runs a Monte Carlo simulation against a causality graph.",
        "operationId": "simulate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Simulation succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or simulation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BindingRequest": {
        "type": "object",
        "description": "A single input-to-distribution binding.",
        "required": [
          "node",
          "distribution"
        ],
        "properties": {
          "distribution": {
            "$ref": "#/components/schemas/DistributionRequest",
            "description": "Distribution to sample from."
          },
          "node": {
            "type": "string",
            "description": "Name of the input node."
          }
        }
      },
      "DistributionRequest": {
        "oneOf": [
          {
            "type": "object",
            "description": "Uniform distribution over `[low, high]`.",
            "required": [
              "low",
              "high",
              "kind"
            ],
            "properties": {
              "high": {
                "type": "number",
                "format": "double",
                "description": "Upper bound."
              },
              "kind": {
                "type": "string",
                "enum": [
                  "uniform"
                ]
              },
              "low": {
                "type": "number",
                "format": "double",
                "description": "Lower bound."
              }
            }
          },
          {
            "type": "object",
            "description": "Normal distribution with mean and standard deviation.",
            "required": [
              "mean",
              "std_dev",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "normal"
                ]
              },
              "mean": {
                "type": "number",
                "format": "double",
                "description": "Mean."
              },
              "std_dev": {
                "type": "number",
                "format": "double",
                "description": "Standard deviation."
              }
            }
          },
          {
            "type": "object",
            "description": "Fixed value.",
            "required": [
              "value",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "fixed"
                ]
              },
              "value": {
                "type": "number",
                "format": "double",
                "description": "Constant value."
              }
            }
          }
        ],
        "description": "A distribution request used by the simulator."
      },
      "EdgeRequest": {
        "type": "object",
        "description": "A directed edge between two nodes.",
        "required": [
          "dependency",
          "dependent"
        ],
        "properties": {
          "dependency": {
            "type": "string",
            "description": "Name of the dependency node."
          },
          "dependent": {
            "type": "string",
            "description": "Name of the dependent node."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Error response body returned for invalid requests.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
      "EvaluateRequest": {
        "type": "object",
        "description": "Request body for `POST /evaluate`.",
        "required": [
          "nodes",
          "edges",
          "inputs"
        ],
        "properties": {
          "edges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EdgeRequest"
            },
            "description": "Directed edges between nodes."
          },
          "inputs": {
            "type": "object",
            "description": "Values for every input node, keyed by node name.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NodeRequest"
            },
            "description": "Nodes that make up the DAG."
          }
        }
      },
      "EvaluateResponse": {
        "type": "object",
        "description": "Response body for `POST /evaluate`.",
        "required": [
          "outputs"
        ],
        "properties": {
          "outputs": {
            "type": "object",
            "description": "Computed value for every node, keyed by node name.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
...
      "NodeRequest": {
        "oneOf": [
          {
            "type": "object",
            "description": "A raw numeric input provided by the caller.",
            "required": [
              "input"
            ],
            "properties": {
              "input": {
                "type": "object",
                "description": "A raw numeric input provided by the caller.",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Human-readable name for the node."
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "description": "A formula from the CASIROS catalog.",
            "required": [
              "formula"
            ],
            "properties": {
              "formula": {
                "type": "object",
                "description": "A formula from the CASIROS catalog.",
                "required": [
                  "name",
                  "kind"
                ],
                "properties": {
                  "kind": {
                    "$ref": "#/components/schemas/FormulaRequest",
                    "description": "The specific formula and its port bindings."
                  },
                  "name": {
                    "type": "string",
                    "description": "Human-readable name for the node."
                  }
                }
              }
            }
          }
        ],
        "description": "A single node in a DAG request."
      },
...
      "SimulateRequest": {
        "type": "object",
        "description": "Request body for `POST /simulate`.",
        "required": [
          "nodes",
          "edges",
          "bindings",
          "target",
          "universe_count"
        ],
        "properties": {
          "bindings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BindingRequest"
            },
            "description": "Input nodes to perturb, each with a distribution."
          },
          "edges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EdgeRequest"
            },
            "description": "Directed edges between nodes."
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NodeRequest"
            },
            "description": "Nodes that make up the DAG."
          },
          "seed": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Optional RNG seed for reproducibility.",
            "minimum": 0
          },
          "target": {
            "type": "string",
            "description": "Name of the node whose output should be aggregated."
          },
          "universe_count": {
            "type": "integer",
            "description": "Number of universes to simulate.",
            "minimum": 0
          }
        }
      },
      "SimulateResponse": {
        "type": "object",
        "description": "Response body for `POST /simulate`.",
        "required": [
          "count",
          "mean",
          "median",
          "min",
          "max"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "description": "Number of universes simulated.",
            "minimum": 0
          },
          "max": {
            "type": "string",
            "description": "Maximum observed value."
          },
          "mean": {
            "type": "string",
            "description": "Mean of the target node."
          },
          "median": {
            "type": "string",
            "description": "Median of the target node."
          },
          "min": {
            "type": "string",
            "description": "Minimum observed value."
          }
        }
      }
    }
  }
}